Public Documentation

IERS Models

This is a list of the supported IERS models and their approximations that can be used to selected the algorithm associated to a specific IERS convention.

2010 Conventions

IERSConventions.iers2010a — Constant

iers2010a

The singleton instance of type IERS2010A, representing the IERS 2010A family of models.

IERSConventions.iers2010b — Constant

iers2010b

The singleton instance of type IERS2010B, representing the IERS 2010B family of models.

Note

This is not an official IERS model.

IERSConventions.CPNc — Constant

CPNc

The singleton instance of type CPNC, representing the concise CPNc from Capitaine & Wallace, Concise CIO based precession-nutation formulations, (2008). This model truncates the X, Y series to deliver an accuracy of few mas.

Note

This is not an official IERS model.

IERSConventions.CPNd — Constant

CPNd

The singleton instance of type CPND, representing the concise CPNd from Capitaine & Wallace, Concise CIO based precession-nutation formulations, (2008). This model truncates the X, Y series to deliver an accuracy of few arcseconds.

Note

This is not an official IERS model.

2003 Conventions

IERSConventions.iers2003a — Constant

iers2003a

The singleton instance of type IERS2003A, representing the IERS 2003A family of models.

IERSConventions.iers2003b — Constant

iers2003b

The singleton instance of type IERS2003B, representing the IERS 2003B family of models.

1996 Conventions

IERSConventions.iers1996 — Constant

iers1996

The singleton instance of type IERS1996, representing the IERS 1996 family of models.

Rotations

CIO-based rotations

IERSConventions.iers_rot3_gcrf_to_cirf — Function

iers_rot3_gcrf_to_cirf(t::Number, m::IERSModel=iers2010b)

Compute the rotation matrix from the Geocentric Celestial Reference Frame (GCRF) to the Celestial Intermediate Reference Frame (CIRF) at time t, expressed in TT seconds since J2000, following the IERS Conventions m.

Note

EOP corrections to the CIP coordinates (δX, δY) are only added in the iers2003a and iers2010a models.

References

IERS Technical Note No. 36

See also

See also iers_rot3_gcrf_to_tirf and iers_rot3_gcrf_to_itrf.

IERSConventions.iers_rot3_gcrf_to_itrf — Function

iers_rot3_gcrf_to_itrf(t::Number, m::IERSModel=iers2010b)

Compute the rotation matrix from the Geocentric Celestial Reference Frame (GCRF) to the International Terrestrial Reference Frame (ITRF) at time t, expressed in TT seconds since J2000, following the IERS Conventions m.

Note

EOP corrections to the CIP coordinates (δX, δY) are only added in the iers2003a and iers2010a models.

Note

Polar motion is neglected in the CPNd model.

References

IERS Technical Note No. 36

See also

See also iers_rot3_gcrf_to_cirf and iers_rot3_gcrf_to_tirf.

IERSConventions.iers_rot3_gcrf_to_tirf — Function

iers_rot3_gcrf_to_tirf(t::Number, m::IERSModel=iers2010b)

Compute the rotation matrix from the Geocentric Celestial Reference Frame (GCRF) to the Terrestrial Intermediate Reference Frame (TIRF) at time t, expressed in TT seconds since J2000, following the IERS Conventions m.

Note

EOP corrections to the CIP coordinates (δX, δY) are only added in the iers2003a and iers2010a models.

References

IERS Technical Note No. 36

See also

See also iers_rot3_gcrf_to_cirf and iers_rot3_gcrf_to_itrf.

IERSConventions.iers_rot3_itrf_to_cirf — Function

iers_rot3_itrf_to_cirf(t::Number, m::IERSModel=iers2010b)

Compute the rotation matrix from the International Terrestrial Reference Frame (ITRF) to the Celestial Intermediate Reference Frame (CIRF) at time t, expressed in TT seconds since J2000, following the IERS Conventions m.

Note

Polar motion is neglected in the CPNd model.

References

IERS Technical Note No. 36

See also

See also iers_rot3_gcrf_to_cirf and iers_rot3_itrf_to_tirf.

IERSConventions.iers_rot3_itrf_to_tirf — Function

iers_rot3_itrf_to_tirf(t::Number, m::IERSModel=iers2010b)

Compute the rotation matrix from the International Terrestrial Reference Frame (ITRF) to the Terrestrial Intermediate Reference Frame (TIRF) at time t, expressed in TT seconds since J2000, following the IERS Conventions m.

Note

Polar motion is neglected in the CPNd model.

References

IERS Technical Note No. 36

See also

See also iers_rot3_gcrf_to_tirf and iers_rot3_itrf_to_cirf.

Equinox-based rotations

IERSConventions.iers_rot3_gcrf_to_gtod — Function

iers_rot3_gcrf_to_gtod(t::Number, m::IERSModel=iers2010b)

Compute the rotation matrix from the Geocentric Celestial Reference Frame (GCRF) to the Greenwich True-of-Date (GTOD) at time t, expressed in TT seconds since J2000.

Note

If the iers1996 conventions are used, the rotation is actually computed starting from the MEME2000 rather than the GCRF.

Note

The EOP nutation corrections are only used for the iers2003a and iers2010a models.

See also

See also iers_rot3_itrf_to_gtod.

IERSConventions.iers_rot3_gcrf_to_mod — Function

iers_rot3_gcrf_to_mod(t::Number, m::IERSModel=iers2010b)

Compute the rotation matrix from the Geocentric Celestial Reference Frame (GCRF) to the Mean-of-Date (MOD) at time t, expressed in TT seconds since J2000.

Note

The Mean-of-Date axes are obtained by applying the frame bias and precession matrix. For this reason, if the iers1996 conventions are used, the rotation is actually computed starting from the MEME2000 rather than the GCRF.

See also

See also iers_pb and iers_rot3_itrf_to_mod.

IERSConventions.iers_rot3_gcrf_to_pef — Function

iers_rot3_gcrf_to_pef(t::Number, m::IERSModel=iers2010b)

Compute the rotation matrix from the Geocentric Celestial Reference Frame (GCRF) to the Pseudo-Earth Fixed (PEF) at time t, expressed in TT seconds since J2000.

Note

If the iers1996 conventions are used, the rotation is actually computed starting from the MEME2000 rather than the GCRF.

Note

The EOP nutation corrections are only used for the iers2003a and iers2010a models.

Note

For the iers1996 and CPNd models, there are no differences between the PEF and GTOD axes. # TODO: specify the magnitude of the rotation!

See also

See also iers_rot3_gcrf_to_gtod and iers_rot3_itrf_to_pef.

IERSConventions.iers_rot3_gcrf_to_tod — Function

iers_rot3_gcrf_to_tod(t::Number, m::IERSModel=iers2010b)

Compute the rotation matrix from the Geocentric Celestial Reference Frame (GCRF) to the True-of-Date (TOD) at time t, expressed in TT seconds since J2000.

Note

The True-of-Date axes are obtained by applying the frame bias, precession and nutation matrix. For this reason, if the iers1996 conventions are used, the rotation is actually computed starting from the MEME2000 rather than the GCRF.

Note

The EOP nutation corrections are only used for the iers2003a and iers2010a models.

See also

See also iers_npb and iers_rot3_itrf_to_tod.

IERSConventions.iers_rot3_itrf_to_mod — Function

iers_rot3_itrf_to_mod(t::Number, m::IERSModel=iers2010b)

Compute the rotation matrix from the International Terrestrial Reference Frame (ITRF) to the Mean-of-Date (MOD) at time t, expressed in TT seconds since J2000.

Note

The EOP nutation corrections are only used for the iers2003a and iers2010a models.

See also

See also iers_rot3_itrf_to_tod and iers_rot3_gcrf_to_mod.

IERSConventions.iers_rot3_itrf_to_tod — Function

iers_rot3_itrf_to_tod(t::Number, m::IERSModel=iers2010b)

Compute the rotation matrix from the International Terrestrial Reference Frame (ITRF) to the True-of-Date (TOD) at time t, expressed in TT seconds since J2000.

Note

The EOP corrections in longitude are only used for the iers2003a and iers2010a models.

See also

See also iers_rot3_itrf_to_gtod and iers_rot3_gcrf_to_tod.

IERSConventions.iers_rot3_itrf_to_gtod — Function

iers_rot3_itrf_to_gtod(t::Number, m::IERSModel=iers2010b)

Compute the rotation matrix from the International Terrestrial Reference Frame (ITRF) to the Greenwich True-of-Date (GTOD) at time t, expressed in TT seconds since J2000.

Note

The CPNd model returns an identity rotation because it neglects the effects of polar motion. # TODO: specify magnitude!

Note

For the iers1996 model, the GTOD and PEF axes are equal because the IERS 1996 conventions do not account for the TIO locator effects.

See also

See also iers_rot3_itrf_to_pef and iers_rot3_gcrf_to_gtod.

IERSConventions.iers_rot3_itrf_to_pef — Function

iers_rot3_itrf_to_pef(t::Number, m::IERSModel=iers2010b)

Compute the rotation matrix from the International Terrestrial Reference Frame (ITRF) to the Pseudo-Earth Fixed (PEF) at time t, expressed in TT seconds since J2000.

Note

The CPNd model returns an identity rotation because it neglects the effects of polar motion. # TODO: specify magnitude!

See also

See also iers_rot3_gcrf_to_pef.

Bias, Precession and Nutation

IERSConventions.iers_bias — Function

iers_bias(m::IERSModel, t::Number)

Compute the frame bias matrix, which transform vectors from the GCRF axes to the Mean Equinox and Mean Equator of J2000 (MEME2000) axes.

Note

Since in the IERS 1996 conventions the bias matrix was still undefined, the returned matrix is the identity.

References

IERS Technical Note No. 21

IERS Technical Note No. 32

IERS Technical Note No. 36

See also

See also iers_precession, iers_pb and iers_npb.

IERSConventions.iers_nutation — Function

iers_nutation(m::IERSModel, t::Number, δΔψ::Number=0, δΔϵ::Number=0)

Compute the nutation matrix that rotates a vector from Mean-of-Date (MOD) to True-of-Date (TOD) axes following the IERS convention m, at time t expressed in TT Julian Centuries since J2000.

Optional EOP nutation corrections can be provided via the δΔψ and δΔϵ parameters.

References

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

See also

See also iers_nutation_comp and iers_obliquity.

IERSConventions.iers_nutation_comp — Function

iers_nutation_comp(m::IERSModel, t::Number)

Compute the nutation components in longitude and obliquity for the IERS convention m, in radians, at time t expressed in TT Julian Centuries since J2000.

Note

For the IAU 2006A model, the function strictly follows the SOFA implementation. It first computes the IAU 2000A nutation, then applies adjustments for the consequences of the change in obliquity from the IAU 1980 ecliptic to the IAU 2006 ecliptic and (ii) for the secular variation in the Earth's dynamical form factor J2. These corrections ensure that the IAU 2000A nutation is consistent with the IAU 2006 precession model. Please note that the coefficients available on the IERS tables already include those corrections, and are retrieved by multiplying the amplitudes of the SOFA nutation in longitude coefficients by 1.00000047.

Note

The expressions of these components for the CPNc and CPNd models are indirectly computed from their CIP series expansion.

Warning

The computation of the free-core nutation and time dependent effects are excluded from this model. To achieve the < 1μas accuracy with the IAU 2006/2000 A precession-nutation models, such effects must be included a-posteriori (through δΔψ and δΔϵ) using the IERS EOP data.

References

IERS Technical Note No. 21

IERS Technical Note No. 32

IERS Technical Note No. 36

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

Capitaine N. and Wallace P. T. (2008), Concise CIO based precession-nutation formulations

ERFA nut06a and nut00b functions

See also

See also iers_nutation

IERSConventions.iers_obliquity — Function

iers_obliquity(m::IERSModel, t::Number)

Compute the mean obliquity of the ecliptic at epoch, in radians, at time t expressed in TT Julian centuries since J2000, according to the IERS convention m.

Note

The mean obliquity for the IERS 2003 conventions already accounts the adjustment to the IAU 1976 precession model for the IAU 2000 precession-rate of the equator in obliquity .

References

IERS Technical Note No. 21

IERS Technical Note No. 32

IERS Technical Note No. 36

IERSConventions.iers_precession — Function

iers_precession(m::IERSModel, t::Number)

Return the precession matrix that rotates a vector from MEME2000 axes to Mean of Date (MOD) axes, at time t expressed in TT Julian centuries since J2000, according to the IERS convention m.

Note

This matrix rotates vectors from the Mean Equator and Mean Equinox of J2000 (MEME2000) to Mean-of-Date (MOD) axes. The frame bias between the GCRF and MEME2000 is excluded from the returned matrix and must eventually be included with a separate rotation.

Note

The IAU Working Group on Precession and the Ecliptic (Hilton, 2006) has decided to leave the choice of the parameterization for the precession angles to the user. In this function for the IERS 1996 conventions, the precession matrix is computed using the traditional parameterization of Newcomb and Liekse (zₐ, θₐ, ζₐ), whereas it adopts the 4-angles formulation (ϵ₀, ψₐ, ωₐ, χₐ) recommended by (Capitaine et al., 2003a) for all the remaining models.

References

Lieske J. H. et al, (1977), Expression for the Precession Quantities Based upon the IAU (1976) System of Astronomical Constants.

Hilton J. L. et al., (2006), Report of the International Astronomical Union Division I Working Group on Precession and the Ecliptic.

Capitaine N. et al., (2003a), Expressions for IAU 2000 precession quantities.

IERS Technical Note No. 21

IERS Technical Note No. 32

IERS Technical Note No. 36

See also

See also precession_angles_rot3 and precession_angles_rot4

IERSConventions.iers_pb — Function

iers_pb(m::IERSModel, t::Number)

Compute the precession-bias (PB) matrix, which transforms vectors from the GCRF axes to Mean-of-Date (MOD) axes, at time t expressed in TT Julian centuries since J2000, according to the IERS convention m

References

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

IERS Technical Note No. 21

IERS Technical Note No. 32

IERS Technical Note No. 36

See also

See also iers_bias, iers_precession and iers_npb.

IERSConventions.iers_npb — Function

iers_npb(m::IERSModel, t::Number, δΔψ=0, δΔϵ=0)

Compute the nutation-bias-precession (NPB) matrix, which transforms vectors from the GCRF to True-of-Date (TOD) axes, at time t expressed in TT Julian centuries since J2000, according to the IERS convention m

References

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

IERS Technical Note No. 21

IERS Technical Note No. 32

IERS Technical Note No. 36

See also

See also iers_nutation, iers_precession and iers_pb.

Celestial Intermediate Pole

IERSConventions.iers_cip_motion — Function

iers_cip_motion(m::IERSModel, t::Number, δX::Number=0, δY::Number=0)

Compute the GCRF-to-CIRF rotation matrix, following the IERS Conventions m, at time t expressed in TT Julian centuries since J2000. Optional IERS EOP corrections for free-core nutation and time dependent effects can be provided via δX and δY

References

IERS Technical Note No. 36

See also

See also cip_xy and cip_xys.

IERSConventions.cip_xy — Function

cip_xy(m::IERSModel, t::Number)

Compute the CIP X and Y coordinates, in radians, following the IERS Conventions m, at time t, expressed in TT Julian centuries since J2000.

Warning

The computation of the free-core nutation and time dependent effects are excluded from this model. To achieve the < 1μas accuracy with the IAU 2006/2000 A precession-nutation models, such effects must be included a-posteriori (through δX and δY) using the IERS EOP data.

References

Capitaine N. and Wallace P. T. (2008), Concise CIO based precession-nutation formulations.

IERS Technical Note No. 21

IERS Technical Note No. 32

IERS Technical Note No. 36

See also

See also iers_cip_motion and cip_xys.

IERSConventions.cip_xys — Function

cip_xys(m::IERSModel, t::Number, δX::Number=0, δY::Number=0)

Compute the CIP X, Y and CIO locator s coordinates, in radians, following the IERS conventions m at time t, expressed in TT Julian centuries since J2000. Optional EOP nutation corrections can be provided via the δX and δY parameters.

Note

Because of the small values of the CIP corrections, the CIO locator holds pretty much irrespective on them. Indeed, some reports compute s with the corrected CIP coordinates, some do not.

References

Capitaine N. and Wallace P. T. (2008), Concise CIO based precession-nutation formulations

IERS Technical Note No. 21

IERS Technical Note No. 32

IERS Technical Note No. 36

See also

See also iers_cip_motion and cip_xy.

IERSConventions.cip_vector — Function

cip_vector(m::IERSModel, t::Number)

Compute the Celestial Intermediate Pole (CIP) vector, following the IERS Conventions m at time t, expressed in TT Julian centuries since J2000.

References

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

See also

See also cip_xy.

Earth Rotation and Sidereal Time

IERSConventions.iers_era — Function

iers_era(m::IERSModel, t::Number)

Compute the Earth Rotation Angle (ERA), in radians, at time t expressed as UT1 days since J2000, according to the IERS convention m.

Note

In the IERS 1996 conventions, θ is referred to as the Stellar Angle.

References

IERS Technical Note No. 21

IERS Technical Note No. 36

See also

See also iers_era_rotm.

IERSConventions.iers_era_rotm — Function

iers_era_rotm(m::IERSModel, t::Number)

Compute the CIRF-to-TIRF rotation matrix, according to the IERS conventions m, at time t expressed in UT1 days since J2000

References

IERS Technical Note No. 36

See also

See also iers_era.

IERSConventions.iers_gmst — Function

iers_gmst(m::IERSModel, t::Number)

Compute the Greenwich Mean Sidereal Time (GMST), in radians, following the IERS Conventions m at time t expressed as TT Julian centuries since J2000.

Note

The input time is automatically converted to UT1 for the computation of the Earth Rotation Angle (ERA) or for the computation of the GMST of the 1996 conventions. Thus, EOP data must be loaded before calling this function.

References

IERS Technical Note No. 21

IERS Technical Note No. 32

IERS Technical Note No. 36

See also

See also iers_gast and iers_era.

IERSConventions.iers_gast — Function

iers_gast(m::IERSModel, t::Number, δΔψ::Number=0)

Compute the Greenwich Apparent Sidereal Time (GAST), in radians, following the IERS Conventions m at time t expressed as TT Julian centuries since J2000. An optional EOP correction for the nutation in longitude can be passed via δΔψ for the exact computation of the equation of the equinoxes.

Note

The input time is automatically converted to UT1 for the computation of the Earth Rotation Angle (ERA) or for the computation of the GMST of the 1996 conventions. Thus, EOP data must be loaded before calling this function.

References

IERS Technical Note No. 21

IERS Technical Note No. 32

IERS Technical Note No. 36

See also

See also iers_gmst, equation_equinoxes and iers_era.

Polar Motion

IERSConventions.iers_polar_motion — Function

iers_polar_motion(m::IERSModel, xₚ::Number, yₚ::Number, t::Number)

Compute the Polar Motion TIRF-to-ITRF rotation matrix, according to the IERS Conventions m, at time t expressed in TT Julian centuries since J2000. The function requires xp and yp, the Celestial Intermediate Pole (CIP) coordinates with respect to the International Celestial Reference Frame (ITFR).

TODO: expand description! (and check assumption for CPNd)

References

IERS Technical Note No. 21

IERS Technical Note No. 32

IERS Technical Note No. 36

EOP Data

IERSConventions.eop_filename — Function

eop_filename(eop::EOPData)

Get the loaded Earth Orientation Parameters (EOP) filename.

IERSConventions.eop_load_data! — Function

eop_load_data!(filename, m::IERS2010)

Initialise the Earth Orientation Parameters (EOP) from a dedicated JSMD .eop.dat file.

Note

Currently, only EOP data associated to the IAU2006/2000A precession-nutation model is supported.

IERSConventions.eop_generate_from_csv — Function

eop_generate_from_csv(m::IERSModel, inputfile, outputfile)

Parse CSV files containing IERS EOP data and extract the relevant information to a dedicated JSMD .eop.dat file. Supported formats are the EOP C04 series and the Rapid Data prediction (finals).

Note

The outputfile name should not include the file extension, which is automatically added by this function.

Note

Depending on the type of file, either the CIP or the nutation corrections may be present. This function applies a conversion algorithm to automatically retrieve the missing data.

Warning

The rapid data prediction files store the LOD, CIP and nutation corrections in milliseconds and milliarcseconds, respectively, whereas the EOP C04 files do not. This routine automatically detects the correct unit of measure by analysing the format of the input filename. Thus, the filename should be kept equal to the one released by the IERS.

References

USNO finals

USNO finals 2000A

See also

See also eop_generate_from_txt.

IERSConventions.eop_generate_from_txt — Function

eop_generate_from_txt(m::IERSModel, inputfile, outputfile)

Parse TXT files containing IERS EOP data and extract the relevant information to a dedicated JSMD .eop.dat file. Supported formats are the EOP C04 series and the Rapid Data prediction (finals).

Note

The outputfile name should not include the file extension, which is automatically added by this function.

Note

Depending on the type of file, either the CIP or the nutation corrections may be present. This function applies a conversion algorithm to automatically retrieve the missing data.

Warning

This routine recognises the file structure and column ordering by analysing the input file name, which should be left equal to the one retrieved from the IERS website.

References

USNO finals

USNO finals 2000A

See also

See also eop_generate_from_csv.