Fundamental Arguments

Delaunay's Arguments

IERSConventions.DelaunayArgs — Type

DelaunayArgs

Container holding the Delaunay arguments of the Nutation Theory.

Fields

M – Mean anomaly of the Moon, in radians.

S – Mean anomaly of the Sun, in radians.

F – Mean longitude of the Moon minus mean longitude of the ascending node, in radians.

D – Mean elongation of the Moon from the Sun, in radians.

Ω – Mean longitude of the Moon's ascending node, in radians.

IERSConventions.delaunay_anomaly_moon — Function

delaunay_anomaly_moon(m::IERSModel, t::Number)

Compute the mean anomaly of the Moon, in radians, given time t expressed in TDB Julian centuries since J2000.

IERSConventions.delaunay_anomaly_sun — Function

delaunay_anomaly_sun(m::IERSModel, t::Number)

Compute the mean anomaly of the Sun, in radians, given time t expressed in TDB Julian centuries since J2000.

IERSConventions.delaunay_longitude_diff — Function

delaunay_longitude_diff(m::IERSModel, t::Number)

Compute the difference between the longitude of the Moon and the longitude of the Moon's node, in radians, given time t expressed in TDB Julian centuries since J2000.

IERSConventions.delaunay_elongation_moon — Function

delaunay_elongation_moon(m::IERSModel, t::Number)

Compute the mean elongation of the Moon from the Sun, in radians, given time t expressed in TDB Julian centuries since J2000.

IERSConventions.delaunay_longitude_node — Function

delaunay_longitude_node(m::IERSModel, t::Number)

Compute the longitude of the mean ascending node of the lunar orbit on th ecliptic, measured from the mean equinox of date, in radians, given time t expressed in TDB Julian centuries since J2000.

Planetary Arguments

IERSConventions.PlanetaryArgs — Type

PlanetaryArgs

Container holding the planetary arguments of the Nutation Theory.

Note

The default constructors skip their computation when neither the iers2003a and iers2010a models are required. This results in a maximum error of about 1.5 mas between 2000 and 2100.

Fields

λ_Me – Mercury's mean heliocentric longitude, in radians.

λ_Ve – Venus's mean heliocentric longitude, in radians.

λ_Ea – Earth's mean heliocentric longitude, in radians.

λ_Ma – Mars's mean heliocentric longitude, in radians.

λ_Ju – Jupiter's mean heliocentric longitude, in radians.

λ_Sa – Saturn's mean heliocentric longitude, in radians.

λ_Ur – Uranus's mean heliocentric longitude, in radians.

λ_Ne – Neptune's mean heliocentric longitude, in radians.

pₐ – General precession in longitude, in radians.

IERSConventions.pa_mercury — Function

pa_mercury(m::IERSModel, t::Number)

Return the mean heliocentric longitude of Mercury in radians, at time t expressed in TDB Julian centuries since J2000.

Warning

This function has not been implemented for the IERS 1996 conventions.

References

IERS Technical Note No. 21

IERS Technical Note No. 36

ERFA software library

IERSConventions.pa_venus — Function

pa_venus(m::IERSModel, t::Number)

Return the mean heliocentric longitude of Venus in radians, at time t expressed in TDB Julian centuries since J2000.

References

IERS Technical Note No. 21

IERS Technical Note No. 36

ERFA software library

IERSConventions.pa_earth — Function

pa_earth(m::IERSModel, t::Number)

Return the mean heliocentric longitude of Earth in radians, at time t expressed in TDB Julian centuries since J2000.

References

IERS Technical Note No. 21

IERS Technical Note No. 36

ERFA software library

IERSConventions.pa_mars — Function

pa_mars(m::IERSModel, t::Number)

Return the mean heliocentric longitude of Mars in radians, at time t expressed in TDB Julian centuries since J2000.

References

IERS Technical Note No. 21

IERS Technical Note No. 36

ERFA software library

IERSConventions.pa_jupiter — Function

pa_jupiter(m::IERSModel, t::Number)

Return the mean heliocentric longitude of Jupiter in radians, at time t expressed in TDB Julian centuries since J2000.

References

IERS Technical Note No. 21

IERS Technical Note No. 36

ERFA software library

IERSConventions.pa_saturn — Function

pa_saturn(m::IERSModel, t::Number)

Return the mean heliocentric longitude of Saturn in radians, at time t expressed in TDB Julian centuries since J2000.

References

IERS Technical Note No. 21

IERS Technical Note No. 36

ERFA software library

IERSConventions.pa_uranus — Function

pa_uranus(m::IERSModel, t::Number)

Return the mean heliocentric longitude of Uranus in radians, at time t expressed in TDB Julian centuries since J2000.

Warning

This function has not been implemented for the IERS 1996 conventions.

References

IERS Technical Note No. 21

IERS Technical Note No. 36

ERFA software library

IERSConventions.pa_neptune — Function

pa_neptune(m::IERSModel, t::Number)

Return the mean heliocentric longitude of Neptune in radians, at time t expressed in TDB Julian centuries since J2000.

Warning

This function has not been implemented for the IERS 1996 conventions.

References

IERS Technical Note No. 21

IERS Technical Note No. 36

ERFA software library

IERSConventions.pa_precession — Function

pa_precession(m::IERSModel, t::Number)

Return the general accumulated precession in longitude pₐ in radians, at time t expressed in TDB Julian centuries since J2000.

References

IERS Technical Note No. 21

IERS Technical Note No. 36

ERFA software library

Fukushima-Williams

IERSConventions.fw2xy — Function

fw2xy(ϵ::Number, ψ::Number, γ::Number, φ::Number)

Compute the CIP X and Y coordinates from Fukushima-Williams bias-precession-nutation angles, in radians.

References

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

IERSConventions.fw_angles — Function

fw_angles(m::IERS2010, t::Number)

Compute the precession angles, γ, ϕ, ψ, ϵ in radians, according to the IAU 2006 Fukushima-Williams 4-angle formulation at time t expressed in TT Julian centuries since J2000.

References

Luzum, B. and Petit G. (2012), The IERS Conventions (2010), IERS Technical Note No. 36

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

ERFA library

IERSConventions.fw_matrix — Function

fw_matrix(γ, ϕ, ψ, ε)

Form a rotation matrix given the Fukushima-Williams angles, expressed in radians.

The present function can construct three different matrices depending on which angles are supplied as arguments:

NPB: To obtain the Nutation-Precession-Bias (NPB) matrix, generate the four standard FW precession angles (̄γ, ̄ϕ, ̄ψ, ϵₐ) then generate the nutation components Δψ and Δϵ and add them to ̄ψ, ϵₐ. Finally, call the present functions using those four angles as arguments.

PB: To obtain the Precession-Bias matrix, generate the four standard FW precession angles and call the present function.

B: To obtain the frame bias matrix, generate the four standard FW precession angles at date J2000.0 and call this function.

The remaining nutation-only and precession only matrices can be obtained by combining these three appropriately.

References

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

ERFA library

Earth Orientation Parameters

Constants

IERSConventions.NutationCorrections — Type

NutationCorrections

Container to hold the nutation corrections and celestial pole offsets associated to a given IERS model (e.g., 2000A)

Fields

δX, δY: Celestial pole offsets referred to the model IAU2000A (rad).

δΔψ, δΔϵ: Nutation corrections in longitude and obliquity (rad)

See also

See also EOPData.

IERSConventions.set_nutation_corr! — Function

set_nutation_corr!(eop::EOPData, m::IERSModel, nc::NutationCorrections)

Set the nutation and cip corrections associated to the IERS model m to nc.

IERSConventions.NutCorrectionsInterpolator — Type

NutCorrectionsInterpolator

Container to store the interpolators for the nutation corrections and celestial pole offsets associated to a given IERS model.

See also

See also EOPInterpolator

IERSConventions.EOPData — Type

EOPData{T}

EOP Data container. Data is parameterised by time expressed in Terrestrial Time (TT) days.

Fields

filename : File where the EOP data are stored.

days_UTC: UTC Julian days since J2000.

cent_TT : TT Julian centuries since J2000.

UT1_TT: UT1 minus TT offset, in seconds.

LOD: Length of day offset (s).

xp, yp: Polar motion with respect to the crust (rad).

nut1996, nut2003, nut2006: nutation and CIP corrections.

IERSConventions.EOPInterpolator — Type

EOPInterpolator

Container to store the interpolators for the loaded EOP data.

IERSConventions.IERS_EOP — Constant

IERS_EOP

Earth Orientation Parameters (EOP) interpolators.

See also

See also eop_load_data! and EOPInterpolator.

IERSConventions.IERS_EOP_DATA — Constant

IERS_EOP_DATA

Earth Orientation Parameters Data.

See also

See also EOPData.

Loading

IERSConventions.eop_read_data — Function

eop_read_data(filename)

Read Earth Orientation Parameters (EOP) from a dedicated JSMD .eop.dat file.

Note

JSMD's .eop.dat EOP files are meant to have a fixed structure to ease the retrieval of the relevant EOP data.

IERSConventions.eop_set_data! — Function

eop_set_data!(filename, m::IERSModel)

Set Earth Orientation Parameters (EOP) to be used for frames transformations from a JSMD .eop.dat file, given a reference model m.

Parsing

IERSConventions.eop_write_data — Function

eop_write_data(data, output_filename)

Write the EOP data stored in the matrix data to a dedicated JSMD .eop-dat file.

Note

The output_filename should not include the file extension, which is automatically added by this function.

See also

See also eop_generate_from_csv and eop_generate_from_txt.

IERSConventions.δnut_to_δcip — Function

δnut_to_δcip(m::IERSModel, t::Number, δΔψ::Number, δΔϵ::Number)

Convert nutation corrections in longitude δΔψ and obliquity δΔϵ to CIP corrections δX, δY at time t expressed in TT Julian centuries since J2000 for model m. All input and output EOP are given in radians.

References

IERS Technical Note No. 36

Paris IERS observatory FTP server (UAI2000 package).

See also

See also δcip_to_δnut.

IERSConventions.δcip_to_δnut — Function

δcip_to_δnut(m::IERSModel, t::Number, δX::Number, δY::Number)

Convert CIP corrections δX, δY to nutation corrections in longitude δΔψ and obliquity δΔϵ to at time t expressed in TT Julian centuries since J2000 for model m. All input and output EOP are given in radians.

References

IERS Technical Note No. 36

Paris IERS observatory FTP server (UAI2000 package).

See also

See also δnut_to_δcip.

Interpolation Functions

IERSConventions.eop_δΔψ — Function

eop_δΔψ(m::IERSModel, t::Number)

Interpolate and retrieve the EOP nutation correction in longitude δΔψ, in radians, at time t expressed in TT Julian centuries since J2000 for the IERS convention m.

IERSConventions.eop_δΔϵ — Function

eop_δΔϵ(m::IERSModel, t::Number)

Interpolate and retrieve the EOP nutation correction in obliquity δΔϵ, in radians, at time t expressed in TT Julian centuries since J2000 for the IERS convention m.

IERSConventions.eop_δX — Function

eop_δX(m::IERSModel, t::Number)

Interpolate and retrieve the CIP δX correction, in radians, at time t expressed in TT Julian centuries since J2000 for the IERS convention m.

IERSConventions.eop_δY — Function

eop_δY(m::IERSModel, t::Number)

Interpolate and retrieve the CIP δY correction, in radians, at time t expressed in TT Julian centuries since J2000 for the IERS convention m.

IERSConventions.eop_xp — Function

eop_xp(m::IERSModel, t::Number)

Interpolate and retrieve the pole xₚ coordinate, in radians, at time t expressed in TT Julian centuries since J2000 for the IERS convention m.

IERSConventions.eop_yp — Function

eop_yp(m::IERSModel, t::Number)

Interpolate and retrieve the pole yₚ coordinate, in radians, at time t expressed in TT Julian centuries since J2000 for the IERS convention m.

IERSConventions.offset_tt2ut1 — Function

offset_tt2ut1(t::Number)

Return the TT-to-UT1 offset, in seconds, at t expressed in TT seconds since J2000.

Poisson Series

IERSConventions.PoissonSeries — Type

PoissonSeries

This container stores the non-polynomial part of a Poisson series, i.e., the sine and cosine coefficients and the Delaunay's and Planetary arguments of the nutation theory multipliers.

IERSConventions.NutationDataParser — Type

NutationDataParser <: PoissonDataParser

Parser for nutation-like data, in which the orders of the different coefficients are stored on separate columns, e.g., in and out of phase.

Fields

ncols: number of columns

c_dargs: first column of the Delaunay's arguments

c_pargs: first column of the planetary arguments

c_sin: column with the sine terms

c_cos: column with the cosine terms

linePattern: regex expression to parse the relevant data lines

IERSConventions.CIODataParser — Type

CIODataParser <: PoissonDataParser

Parser for CIO-like data, in which the orders of the different coefficients are separated by a specific line header in the form of " j = 0 Nb of terms = 1306".

Fields

ncols: number of columns

c_dargs: first column of the Delaunay's arguments

c_pargs: first column of the planetary arguments

c_sin: column with the sine terms

c_cos: column with the cosine terms

linePattern: regex expression to parse the relevant data lines

IERSConventions.parse_iers_constants — Function

parse_iers_constants(filename::String, parser::PoissonDataParser)

Retrieve the Poisson series within filename with the configuration of parser.

IERSConventions.generate_iers_file — Function

generate_iers_file(filename::String, constant_name::Symbol, constant_data)

Generate or append the data in constant_data to filename as a variable named constant_name.

Miscellaneous

IERSConventions.cio_locator — Function

cio_locator(m::IERSModel, t::Number, x::Number, y::Number)

Compute the CIO locator s, in radians, following the IERS Conventions m' at timet, expressed inTTJulian centuries since J2000, given the CIP coordinatesxandy`.

References

Capitaine N. and Wallace P. T. (2008), Concise CIO based precession-nutation formulations

IERS Technical Note No. 21

IERS Technical Note No. 32

IERS Technical Note No. 36

See also

See also iers_cip_motion, cip_xy and cip_xys.

IERSConventions.equation_equinoxes — Function

equation_equinoxes(m::IERSModel, t::Number, δΔψ::Number = 0)

Compute the Equation of the Equinoxes, in radians, according to the IERS Conventions m, at time t expressed as TT Julian centuries since J2000. An optional EOP correction for the nutation in longitude can be passed via δΔψ.

Note

In this framework, the equation of the equinoxes is always defined according to the following equation: GAST = GMST + equation of the equinoxes

References

IERS Technical Note No. 21

IERS Technical Note No. 32

IERS Technical Note No. 36

See also

See also iers_obliquity, iers_nutation_comp and eeq_complementary.

IERSConventions.eeq_complementary — Function

eeq_complementary(m::IERSModel, t::Number)

Compute the complementary terms of the equation of the equinoxes, in radians, associated to the IERS Conventions m, at time t expressed in TT Julian centuries since J2000.

Note

For the IERS 1996 Conventions, starting from 1997-02-27T00:00:00 UTC, two additional terms are included to account for the Moon effect (see IAU 1997 C7 resolution). The complementary terms of earlier dates are always null.

References

IAU Resolution C7, Recommendation 3 (1994)

IERS Technical Note No. 21

IERS Technical Note No. 32

IERS Technical Note No. 36

IERSConventions.npb2xy — Function

npb2xy(A::AbstractMatrix)

Retrieve the CIP X and Y coordinates from the nutation-precession-bias matrix, in radians.

References

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

IERSConventions.precession_angles_rot3 — Function

precession_angles_rot3(m::IERSModel, t::Number)

Return the equatorial precession angles zₐ, θₐ, ζₐ, in radians, at time t expressed in TT Julian centuries since J2000 for the 3-rotations precession series initially used by Newcomb and Lieske.

Note

The expressions for these angles compatible with the IAU 2000A precession and nutation have been developed in order to match the 4-rotation series to sub-microarcsecond accuracy over 4 centuries.

Note

TT is used as time argument instead of TDB. The largest term in the TDB-TT difference causes a periodic error which is significantly under the microarcsecond accuracy.

References

Lieske J. H. et al, (1977), Expression for the Precession Quantities Based upon the IAU (1976) System of Astronomical Constants.

Capitained N. 2003b et al., Expressions for IAU 2000 precession quantities.

Hilton J. L. et al., (2006), Report of the International Astronomical Union Division I Working

Group on Precession and the Ecliptic.

IERS Technical Note No. 21

IERS Technical Note No. 32

IERS Technical Note No. 36

See also

See also iers_precession and precession_angles_rot4

IERSConventions.precession_angles_rot4 — Function

precession_angles_rot4(m::IERSModel, t::Number)

Return the precession angles ϵ₀, ψₐ, ωₐ, χₐ, in radians, at time t expressed in TT Julian centuries since J2000 required for the canonical 4-rotations precession series.

Note

TT is used as time argument instead of TDB. The largest term in the TDB-TT difference causes a periodic error which is significantly under the microarcsecond accuracy.

References

IERS Technical Note No. 21

IERS Technical Note No. 32

IERS Technical Note No. 36

See also

See also iers_precession and precession_angles_rot4

IERSConventions.tio_locator — Function

tio_locator(m::IERSModel, t::Number)

Compute the TIO locator s' at date, positioning the Terrestrial Intermediate Origin on the equator of the Celestial Intermediate Pole (CIP) at time t expressed as TT Julian centuries since J2000.

This function approximates the unpredictable motion of the TIO locator s' with its secular drift of ~0.47 μas/century.

Note

A null value is returned for the iers1996 and the CPNd models.

References

IERS Technical Note No. 36

Lambert, S. and Bizouard C. (2002), Positioning the Terrestrial Ephemeris Origin in the Terrestrial Reference Frame, DOI: 10.1051/0004-6361:20021139

IERSConventions.xys2m — Function

xys2m(x::Number, y::Number, s::Number)

Compute the GCRF-to-CIRF matrix given the CIP x, y' coordinates and the CIO locators`, all in radians.

References

Wallace P. T. and Capitaine N. (2006), Precession-nutation procedures consistent with IAU 2006 resolutions, DOI: 10.1051/0004-6361:20065897

ERFA library